added functions to modules
