+++ package.json
@@ -39,6 +39,7 @@
     "stylelint-csstree-validator": "^1.9.0",
     "stylelint-scss": "3.17.x",
     "ts-loader": "^8.1.0",
+    "uuid": "^3.4.0",
     "webpack": "^5.28.0",
     "webpack-cli": "^4.6.0",
     "webpack-dev-server": "^3.11.2"
